// src/app/student/interfaces/student.interfaces.ts
// Generated by https://quicktype.io

export interface SearchStudent {
  data:  Student[];
  meta:  Meta;
  type:  string;
  links: Links;
}

export interface Student {
  id:                            number;
  names:                         string;
  lastnames:                     string;
  dateBirth:                     string;
  nacionality:                   string;
  departmentBirth:               string;
  municipalityBirth:             string;
  yearStudy:                     string;
  nie:                           string;
  departureNumber:               string;
  departureFolio:                string;
  departureBook:                 string;
  anotherIdentificationDocument: string;
  salvadoreno_por:               string;
  incomeSpecialty:               string;
  parvularianStudy:              string;
  repeatSection:                 string;
  providencisShoolCenters:       string;
  previousYear:                  number;
  bloodType:                     string;
  sexo:                          string;
  stateFamiliar:                 string;
  disability:                    string;
  email:                         string;
  telephone:                     string;
  zone:                          string;
  departmentResidence:           string;
  municipalityResidence:         string;
  cantonResidence:               string;
  hamletResidence:               string;
  streetType:                    string;
  fullAddress:                   string;
  middleTransport:               string;
  distanceFromHomeSchool:        string;
  riskFactor:                    string;
  numberMembersFamily:           number;
  integrated:                    string;
  dependsEconomically:           string;
  hasChildren:                   string;
  numberChildren:                string;
  work:                          string;
  coexistenceWith:               string;
  mothersName:                   string;
  occupationMother:              string;
  workplaceOfTheMother:          string;
  mothersPhone:                  string;
  fathersName:                   string;
  occupationFather:              string;
  workplaceOfTheFather:          string;
  fathersPhone:                  string;
  ResponsibleNames:              string;
  ResponsibleLastNames:          string;
  DuiResponsible:                string;
  familyStateResponsible:        string;
  emailResponsible:              string;
  phoneResponsible:              string;
  zoneReponsible:                string;
  departmentReponsible:          string;
  municipalityReponsible:        string;
  cantonReponsible:              string;
  hamletReponsible:              string;
  streetTypeReponsible:          string;
  fullAddressResponsible:        string;
  professionIfficeResponsible:   string;
  responsibleSchool:             string;
  riskFactorResponsable:         string;
  dateReviewForm:                string;
  birthCertificate:              string;
  certificationNotes:            string;
  certificate:                   string;
  photos:                        string;
  RecordNotes:                   string;
  residentCard:                  string;
  teacher:                       string;
  created_at:                    string;
}

export interface Links {
  first: string;
  last:  string;
  prev:  null;
  next:  null;
}

export interface Meta {
  organization: string;
  authors:      string[];
  current_page: number;
  from:         number;
  last_page:    number;
  links:        Link[];
  path:         string;
  per_page:     number;
  to:           number;
  total:        number;
}

export interface Link {
  url:    null | string;
  label:  string;
  active: boolean;
}